Remove unnecessary dtype annotations
This wasn't the reason the tests were failing.

+++ tests/cuthbertlib/kalman/utils.py
@@ -22,13 +22,13 @@ def generate_lgssm(seed: int, x_dim: int, y_dim: int, num_time_steps: int):
     ys = []
 
     key, sub_key = random.split(key)
-    x = m0 + chol_P0 @ random.normal(sub_key, (x_dim,), dtype=jnp.float64)
+    x = m0 + chol_P0 @ random.normal(sub_key, (x_dim,))
 
     for t in range(num_time_steps):
         key, state_key, obs_key = random.split(key, 3)
-        state_noise = chol_Q @ random.normal(state_key, (x_dim,), dtype=jnp.float64)
+        state_noise = chol_Q @ random.normal(state_key, (x_dim,))
         x = F @ x + c + state_noise
-        obs_noise = chol_R @ random.normal(obs_key, (y_dim,), dtype=jnp.float64)
+        obs_noise = chol_R @ random.normal(obs_key, (y_dim,))
         y = H @ x + d + obs_noise
         ys.append(y)
 
@@ -37,14 +37,14 @@ def generate_lgssm(seed: int, x_dim: int, y_dim: int, num_time_steps: int):
 
 
 def generate_cholesky_factor(key: KeyArray, dim: int) -> Array:
-    chol_A = random.uniform(key, (dim, dim), dtype=jnp.float64)
+    chol_A = random.uniform(key, (dim, dim))
     chol_A = chol_A.at[jnp.triu_indices(dim, 1)].set(0.0)
     return chol_A
 
 
 def generate_init_model(key: KeyArray, x_dim: int) -> tuple[Array, Array]:
     keys = random.split(key)
-    m0 = random.normal(keys[0], (x_dim,), dtype=jnp.float64)
+    m0 = random.normal(keys[0], (x_dim,))
     chol_P0 = generate_cholesky_factor(keys[1], x_dim)
     return m0, chol_P0
 
@@ -52,8 +52,8 @@ def generate_init_model(key: KeyArray, x_dim: int) -> tuple[Array, Array]:
 def generate_trans_model(key: KeyArray, x_dim: int) -> tuple[Array, Array, Array]:
     keys = random.split(key, 3)
     # Generate transition matrix F and vector c such that the system is stable
-    F = jnp.eye(x_dim) + 0.1 * random.normal(keys[0], (x_dim, x_dim), dtype=jnp.float64)
-    c = 0.1 * random.normal(keys[1], (x_dim,), dtype=jnp.float64)
+    F = jnp.eye(x_dim) + 0.1 * random.normal(keys[0], (x_dim, x_dim))
+    c = 0.1 * random.normal(keys[1], (x_dim,))
     chol_Q = generate_cholesky_factor(keys[2], x_dim)
     return F, c, chol_Q
 
@@ -62,8 +62,8 @@ def generate_obs_model(
     key: KeyArray, x_dim: int, y_dim: int
 ) -> tuple[Array, Array, Array]:
     keys = random.split(key, 3)
-    H = random.uniform(keys[0], (y_dim, x_dim), dtype=jnp.float64)
-    d = random.uniform(keys[1], (y_dim,), dtype=jnp.float64)
+    H = random.uniform(keys[0], (y_dim, x_dim))
+    d = random.uniform(keys[1], (y_dim,))
     chol_R = generate_cholesky_factor(keys[2], y_dim)
     return H, d, chol_R
